Feat : 메코의 질문 로직 추가
